Fix bug in built-in benchmark fitness functions (#9)
* Fix bug in built-in benchmark fitness functions

The built-in GBO implementations of Sphere and Rosenbrock contained a bug: the vector of dependent subfunctions was initialized to a certain size and then extended with further numbers, leaving the subfunction `0` in the initialized part to always be included in the list of dependent subfunctions. This commit/PR fixes that initialization issue.

The impact of this issue was primarily visible in the evaluation count needed to reach the value-to-reach in those two fitness functions, which tended to be inflated due to this bug.

This commit also makes two other changes:
* It fixes another bug which caused (real-valued) conditional distributions to be discarded. This is done by wrapping a call to `initializeDistributions` in the real-valued `population.cpp` in an `if`-check.
* It introduces a BBO sphere function. Previously, only a GBO version was available of this benchmark function, and since the Rosenbrock function had two versions, and I needed a BBO version of sphere for testing, I thought it would be beneficial to include this, too.

* Undo different formatting

* Fix time condition issue

* Fix distribution multiplier int/double bug

+++ gomea/src/real_valued/linkage_model.hpp
@@ -31,7 +31,7 @@ class linkage_model_rv_t : public linkage_model_t {
 			void initializeDistributions();
 
 			double getAcceptanceRate(); 
-			int getDistributionMultiplier( int element_index );
+			double getDistributionMultiplier( int element_index );
 			
 			void addGroup( vec_t<int> group );
 			void addGroup( distribution_t *dist );
